update tour guide service by adding endpoint for soft delete account and add hiking point expertise

+++ src/main/java/com/abdav/giri_guide/controller/MountainController.java
@@ -49,7 +49,7 @@ public class MountainController {
             @RequestParam String status,
             @RequestParam String message,
             @RequestParam boolean useSimaksi,
-            @RequestParam Integer priceSimaksi,
+            @RequestParam Double priceSimaksi,
             @RequestParam MultipartFile image
 
     ) {
